Dialogue scriptable objects' own file creation
Moved every scriptable object class to its own file. Got rid of the DialogueDataStructure file because of some conflicts in the inspector.
