+++ GOLF/Assets/_Project/Scripts/Dialogue/Dialogue.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using UnityEngine;
+
+namespace Golf
+{
+    [CreateAssetMenu(fileName = "NewDialogue", menuName = "Dialogue System/Dialogue")]
+    public class Dialogue : ScriptableObject
+    {
+        public List<DialogueLine> DialogueLines = new List<DialogueLine>();
+    }
+}
