+++ GOLF/Assets/_Project/Scripts/Dialogue/DialogueCharacter.cs
@@ -1,0 +1,12 @@
+﻿using UnityEngine;
+
+namespace Golf
+{
+    [CreateAssetMenu(fileName = "DialogueCharacter", menuName = "Dialogue System/DialogueCharacter")]
+    public class DialogueCharacter : ScriptableObject
+    {
+        public string Name;
+        public Sprite Icon;
+        public string AudioName;
+    }
+}
