+++ GOLF/Assets/_Project/Scripts/Dialogue/DialogueCharacter.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2b41464446f66c44492d457ea5d10e3a
+guid: c8f2ef42025e8bf4cb0da4e4bbd87ec6
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
