+++ GOLF/Assets/_Project/Scripts/Dialogue/DialogueManager.cs
@@ -103,4 +103,12 @@ namespace Golf
             _dialogueButton.onClick.Invoke();
         }
     }
+
+    [Serializable]
+    public class DialogueLine
+    {
+        public DialogueCharacter Character;
+        [TextArea(3, 10)]
+        public string Line;
+    }
 }
